remove EventJson trait
